feat: add OpportunityBottom component
