+++ packages/apps/crm/src/components/organisms/index.js
@@ -9,3 +9,4 @@ export {default as ProspectBody} from './ProspectBody/ProspectBody';
 export {default as ProspectBottom} from './ProspectBottom/ProspectBottom';
 export {default as OpportunityHeader} from './OpportunityHeader/OpportunityHeader';
 export {default as OpportunityBody} from './OpportunityBody/OpportunityBody';
+export {default as OpportunityBottom} from './OpportunityBottom/OpportunityBottom';
